requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# UI Framework
textual==0.47.1

# Report Generation
reportlab==4.0.7
pandas==2.1.4
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
click==8.1.7